Refactored GUI classes into `.gui` subpackage.
