+++ app/src/main/res/layout/fragment_launcher.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <com.platypus.android.server.SwipeOnlySwitch
+    <com.platypus.android.server.gui.SwipeOnlySwitch
         android:id="@+id/launcher_launch_switch"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
